+++ ui/user/src/lib/services/user/mcp.ts
@@ -1,4 +1,5 @@
 import type { CompositeLaunchFormData } from '$lib/components/mcp/CatalogConfigureForm.svelte';
+import { encodeUtf8ToBase64 } from '$lib/format';
 import { mcpServersAndEntries, profile } from '$lib/stores';
 import { getUserDisplayName } from '$lib/utils';
 import {
@@ -18,6 +19,7 @@ import {
 	type RuntimeFormData,
 	type SystemMCPServerCatalogEntry
 } from '..';
+import { AiClient } from './constants';
 
 export interface MCPServerInfo extends MCPServer {
 	id?: string;
@@ -1044,4 +1046,63 @@ export async function disconnectMcpServerUser(server: MCPCatalogServer): Promise
 		return;
 	}
 	await UserService.deleteSingleOrRemoteMcpServer(server.id);
+}
+
+export function getAiClientCommand(client: AiClient, url: string): string {
+	const commands = {
+		[AiClient.Claude]: `claude mcp add ${url}`,
+		[AiClient.Codex]: `codex mcp add ${url}`
+	};
+	return commands[client as keyof typeof commands] ?? '';
+}
+
+const obotApiKeyInput = {
+	type: 'promptString',
+	id: 'obot-api-key',
+	description: 'Obot API Key',
+	password: true
+} as const;
+
+function toVsCodeMcpServerName(displayName: string, connectUrl: string): string {
+	const fromUrl = connectUrl.match(/\/mcp-connect\/([^/?#]+)/)?.[1];
+	if (fromUrl) {
+		return fromUrl;
+	}
+
+	const sanitized = displayName
+		.trim()
+		.replace(/\s+(.)/g, (_, char: string) => char.toUpperCase())
+		.replace(/[^a-zA-Z0-9]/g, '');
+
+	return sanitized || 'obotMcpServer';
+}
+
+function generateCursorMagicLink(displayName: string, url: string): string {
+	const cursorConfig = {
+		type: 'http',
+		url: url
+	};
+	const cursorBase64 = encodeUtf8ToBase64(JSON.stringify(cursorConfig));
+	return `cursor://anysphere.cursor-deeplink/mcp/install?name=${encodeURIComponent(displayName)}&config=${encodeURIComponent(cursorBase64)}`;
+}
+
+function generateVsCodeMagicLink(displayName: string, url: string): string {
+	const vscodeConfig = {
+		name: toVsCodeMcpServerName(displayName, url),
+		inputs: [obotApiKeyInput],
+		type: 'http',
+		url: url,
+		headers: {
+			Authorization: 'Bearer ${input:obot-api-key}'
+		}
+	};
+	return `vscode:mcp/install?${encodeURIComponent(JSON.stringify(vscodeConfig))}`;
+}
+
+export function getAiClientMagicLink(client: AiClient, displayName: string, url: string): string {
+	const fn = {
+		[AiClient.Cursor]: generateCursorMagicLink,
+		[AiClient.VSCode]: generateVsCodeMagicLink
+	};
+	return fn[client as keyof typeof fn] ? fn[client as keyof typeof fn](displayName, url) : '';
 }
